studies/health-easy/requirements.txt
faiss-cpu
openpyxl
langchain
langchain-openai
langchain-community
langchain_groq
langgraph
python-dotenv
pypdf
streamlit
langchainhub
bs4
chromadb
matplotlib
scikit-learn
pypdf2
fastapi
uvicorn
llama-parse
langserve
sse_starlette